requirements.txt
scrapfly-sdk
streamlit
pandas
altair
supabase
python-dotenv
requests